requirements.txt
aes==1.0.0
asgiref==3.5.2
attrs==21.4.0
bitarray==2.3.4
bitstring==3.1.9
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
distlib==0.3.4
Django==3.2.13
djangorestframework==3.13.1
drf-spectacular==0.22.1
filelock==3.7.1
guppy3==3.0.10.post1
ibm-db==3.1.2
ibm-db-django==1.5.1.0
idna==2.9
importlib-metadata==4.11.4
importlib-resources==5.8.0
inflection==0.5.1
jsonschema==4.6.0
kiwisolver==1.2.0
matplotlib==3.2.1
Naked==0.1.31
numpy==1.18.1
opencv-python==4.2.0.34
pandas==1.0.4
Pillow==7.0.0
pipenv==2022.6.7
platformdirs==2.5.2
pycryptodome==3.10.4
pyglet==1.5.3
pyparsing==2.4.7
Pyqtree==1.0.0
pyrsistent==0.18.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4.1
regex==2022.6.2
requests==2.23.0
scipy==1.4.1
shellescape==3.8.1
six==1.15.0
sqlparse==0.4.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.25.8
virtualenv==20.14.1
virtualenv-clone==0.5.7
zipp==3.8.0
